Apply stricter compiler warnings for gfxrecon_format|util|graphics targets

+++ framework/decode/file_processor.h
@@ -96,7 +96,7 @@ class FileProcessor
 
     const std::vector<format::FileOptionPair>& GetFileOptions() const { return file_options_; }
 
-    uint32_t GetCurrentFrameNumber() const { return current_frame_number_; }
+    uint64_t GetCurrentFrameNumber() const { return current_frame_number_; }
 
     uint64_t GetCurrentBlockIndex() const { return block_index_; }
